+++ R/dynamite.R
@@ -135,6 +135,8 @@
 #' # Please update your rstan and StanHeaders installation before running
 #' # on Windows
 #' if (!identical(.Platform$OS.type, "windows")) {
+#'   threads <- data.table::getDTthreads()
+#'   data.table::setDTthreads(threads = 1) # For CRAN
 #'   fit <- dynamite(
 #'     dformula = obs(y ~ -1 + varying(~x), family = "gaussian") +
 #'       lags(type = "varying") +
@@ -145,6 +147,7 @@
 #'     chains = 1,
 #'     refresh = 0
 #'   )
+#'   data.table::setDTthreads(threads = threads) # Restore
 #' }
 #' }
 #'